Update README.md
I have just gone through the install process, twice, my bad, I am a complete ubuntu, RR, novice.  I make these suggestions for COMPLETE clarity regarding the process and to highlight issues encountered in the last 24hrs.
Can I please request a discussion with the developers regarding some proposed developments we are working on as an adjunct to this wonderful project?
Thank you for an amazing effort.  I hope that we are configured, modified and up and running soon on our local node for a citizen's shadow for The Shire of Broome, Western Australia, Australia. 
Regards
Andrew